Continue work on auth

+++ backend/src/models/id_token.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct IdToken {
     #[serde(rename = "iss")]
     pub issuer: String,
@@ -11,8 +11,8 @@ pub struct IdToken {
     pub audience: String,
     // Seconds since 1970-01-01T00:00:00Z measured in UTC.
     #[serde(rename = "exp")]
-    pub expires_at: i128,
+    pub expires_at: i64,
     // Seconds since 1970-01-01T00:00:00Z measured in UTC.
     #[serde(rename = "iat")]
-    pub issued_at: i128,
+    pub issued_at: i64,
 }
